broke out CollectionReader and CollectionWriter
